test: migrate test files from JavaScript to TypeScript
