+++ package.json
@@ -92,6 +92,7 @@
     "rimraf": "^6.1.3",
     "sinon": "^21.0.1",
     "supertest": "^7.2.2",
+    "ts-jest": "^29.4.6",
     "ts-node": "^10.9.2",
     "tsx": "^4.21.0",
     "typescript": "5.7.3",
